Uid component refactor.
