+++ .vscode/settings.json
@@ -6,5 +6,6 @@
   "prettier.resolveGlobalModules": true,
   "prettier.singleQuote": true,
   "prettier.withNodeModules": true,
-  "editor.formatOnType": true
+  "editor.formatOnType": true,
+  "terminal.integrated.automationProfile.windows": {}
 }
