+++ app/not-found.tsx
@@ -1,5 +1,5 @@
 import UseNavigationButton from '@/src/navigate/provider';
-import styles from './page.module.css'
+import styles from './page.module.css';
 
 export default function NotFound() {
   return (
