+++ app/page.module.css
@@ -1,2 +1,3 @@
 .main {
+  justify-content: center;
 }
