+++ src/loading/fallback.tsx
@@ -6,7 +6,6 @@ export default function LoadingFallback() {
     <main className={styles.main}>
       <h2>Page Loading</h2>
       <p>The page you are looking for is currently loading. Please wait.</p>
-
       <UseNavigationButton route='refresh'></UseNavigationButton>
       <UseNavigationButton route='back'></UseNavigationButton>
       <UseNavigationButton route='home'></UseNavigationButton>
