--- src/user/provider.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-import { ReactNode, useState } from 'react';
-
-import { UidState } from './types';
-import UidContext from './context';
-
-export default function UidProvider({ children }: { children: ReactNode }) {
-  const [uid, setUid] = useState<UidState>({ username: '' });
-
-  return (
-    <UidContext.Provider value={{ ...uid, setUid }}>
-      {children}
-    </UidContext.Provider>
-  );
-}
