--- src/user/types.tsx
@@ -1,11 +1,0 @@
-import { Dispatch, SetStateAction } from 'react';
-
-type UidState = {
-  username: string;
-};
-
-type Uid = UidState & {
-  setUid: Dispatch<SetStateAction<UidState>>;
-};
-
-export type { UidState, Uid };
